Mark "excluded" (non-distinct / non-unique / ambiguous) taxa

Takes as an input data frame with Sample ID, Taxa ID, and phlogenetic name fields and returns a similar dataframe with a column for "excluded" taxa (TRUE or FALSE).

Excluded taxa are refered to by multiple names; ambiguous, non-distinct, and non-unique. The "excluded" name was chosen so as to be consistent with "non-target" taxa. That is, taxa marked as "TRUE" are treated as undesireables. Excluded taxa are those that are present in a sample when taxa of the same group are present in the same sample are identified finer level. That is, the parent is marked as excluded when child taxa are present in the same sample.

markExcluded(
  df_samptax,
  SampID = "SAMPLEID",
  TaxaID = "TAXAID",
  TaxaCount = "N_TAXA",
  Exclude = "EXCLUDED",
  TaxaLevels,
  Exceptions
)

Arguments

df_samptax
Input data frame.

SampID
Column name in df_samptax for sample identifier. Default = "SAMPLEID".

TaxaID
Column name in df_samptax for organism identifier. Default = "TAXAID".

TaxaCount
Column name in df_samptax for organism count. Default = "N_TAXA".

Exclude
Column name for Excluded Taxa results in returned data frame. Default = "Excluded".

TaxaLevels
Column names in df_samptax that for phylogenetic names to be evaluated. Need to be in order from coarse to fine (i.e., Phylum to Species).

Exceptions
NA or two column data frame of synonyms or other exceptions. Column 1 is the name used in the TaxaID column of df_samptax. Column 2 is the name used in the TaxaLevels columns of df_samptax.

Value

Returns a data frame of df_samptax with an additional column, ExcludedTaxa.

Details

The excluded taxa are referenced in the metric values function. These taxa are removed from the taxa richness metrics. This is because these are coarser level taxa.

Exceptions is a 2 column data frame of synonyms or other exceptions. Column 1 is the name used in the TaxaID column the input data frame (df_samptax). Column 2 is the name used in the TaxaLevels columns of the input data frame (df_samptax). The phylogenetic columns (TaxaLevels) will be modified from Column 2 of the Exceptions data frame to match Column 1 of the Exceptions data frame. This ensures that the algorithm for markExcluded works properly. The changes will not be stored and the original names provided in the input data frame (df_samptax) will be returned in the final result. The function example below includes a practical case.

Taxa Levels are phylogenetic names that are to be checked. They should be listed in order from course (kingdom) to fine (species). Names not appearing in the data will be skipped.

The spelling of names must be consistent (including case) for this function to produce the intended output.

Examples