Calculate metric values, Bugs

Subfunction of metric.values for use with Benthic Macroinvertebrates

metric.values.bugs(
  myDF,
  MetricNames = NULL,
  boo.Adjust = FALSE,
  cols2keep = NULL,
  MetricSort = NA,
  boo.marine = FALSE,
  boo.Shiny,
  verbose,
  metric_subset,
  taxaid_dni = NULL
)

Arguments

myDF

Data frame of taxa.

MetricNames

Optional vector of metric names to be returned.

boo.Adjust

Optional boolean value on whether to perform adjustments of values prior to scoring. Default = FALSE but may be TRUE for certain metrics.

cols2keep

Column names of fun.DF to retain in the output. Uses column names.

MetricSort

How metric names should be sort; NA = as is , AZ = alphabetical. Default = NULL.

boo.marine

Should estuary/marine metrics be included. Ignored if fun.MetricNames is not null. Default = FALSE.

boo.Shiny

Boolean value for if the function is accessed via Shiny. Default = FALSE.

verbose

Include messages to track progress. Default = FALSE

metric_subset

Subset of metrics to be generated. Internal function. Default = NULL

taxaid_dni

Taxa names to be included in DNI (Do Not Include) metrics (n = 3) but dropped for all other metrics. Only for benthic metrics. Default = NULL

Value

Data frame

Details

For internal use only. Called from metric.values().